pyproject.toml
[tool.poetry]
name = "module-snapshot"
version = "0.1.0"
description = ""
authors = ["Lucas <lhenriquerod@dxc.com>"]
readme = "README.md"
packages = [{include = "module_snapshot"}]

[tool.poetry.dependencies]
python = "^3.10"
azure-mgmt-compute = "^30.0.0"
azure-identity = "^1.13.0"
dynaconf = "^3.1.12"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pylint = "^2.17.4"
pytest-cov = "^4.1.0"
bandit = "^1.7.5"
safety = "^2.3.5"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
